Refactors test code structure, adds tests, and introduces reqlite
The existing testing approach was adding test data to the production
RethinkDB database tables, which was not desirable.  Additionally,
the test code structure did not align with the code files being.
The new structure is split up to make it easier to see what code
is being tested in each file.